test(l1): move guest tx type tests to test/
