+++ test/Cargo.toml
@@ -26,6 +26,7 @@ ethrex-levm.workspace = true
 ethrex-rpc.workspace = true
 
 [dev-dependencies]
+ethrex-guest-program.workspace = true
 ethrex-vm.workspace = true
 once_cell = "1.21"
 rustc-hash.workspace = true
