+++ test/tests/guest_program/mod.rs
@@ -1,0 +1,1 @@
+mod l1_transaction_types_tests;
